% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confdf.R
\name{confdf}
\alias{confdf}
\title{NBA Winning Percentage Data Frame, by Conference}
\usage{
confdf(conference, year)
}
\arguments{
\item{conference}{\itemize{
\item character value, either 'eastern' or 'western'
}}

\item{year}{\itemize{
\item numeric value, a year in which the desired season has ended (e.g. year = 2008 will provide
a data frame for 2007-2008 season, year = 2017 for 2016-17 etc.). Please note, that the function only accepts
years in between 1977 and 2020.
}}
}
\description{
Allows to create a data frame of NBA teams from a chosen conference, with their winning percentage
}
\examples{
df.west_2007 <- confdf('western', 2007)
df.east_1999 <- confdf('eastern', 1999)
}
