% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confplot.R
\name{confplot}
\alias{confplot}
\title{NBA Conferences Winning Percentage Plotter}
\usage{
confplot(conference, year)
}
\arguments{
\item{conference}{\itemize{
\item character value, either 'eastern' or 'western'
}}

\item{year}{\itemize{
\item numeric value, a year in which the desired season has ended
(e.g. year = 2008 will provide a graph for a 2007-2008 season,
year = 2017 for 2016-17 etc.).
Please note, that the function only accept years fr0m 1970.
}}
}
\description{
Creates a winning percentage bar chart for NBA teams
in a chosen conference and year (from 1977 to 2020)
}
\examples{
confplot('eastern', 2007)
confplot('western', 1999)
}
