% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/divplot.R
\name{divplot}
\alias{divplot}
\title{NBA Divisions Winning Percentage Plotter}
\usage{
divplot(division, first_year, last_year)
}
\arguments{
\item{division}{\itemize{
\item character value, 'northwest', 'pacific', 'southwest'
for the Western Conference, and 'atlantic', 'central', 'southeast'
for the Eastern Conference
}}

\item{first_year}{\itemize{
\item first year, from which the data will be
presented on the graph
}}

\item{last_year}{\itemize{
\item last year, that closes the selected timespan
}}
}
\description{
Creates a winning percentage boxplot for NBA teams
in a chosen division and selected timespan
(between 2005 and 2020)
}
\examples{
divplot('southwest', 2014, 2018)
divplot('central', 2009, 2012)
}
