% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/teamdf.R
\name{teamdf}
\alias{teamdf}
\title{NBA Winning Percentage Data Frame, by Team}
\usage{
teamdf(team, year = 0)
}
\arguments{
\item{team}{\itemize{
\item character value, a team's abbreviation
(e.g. 'BOS' for Boston Celtics)
}}

\item{year}{\itemize{
\item numeric value, a year in which the desired
season has ended (e.g. year = 2008 will provide
a data frame for 2007-2008 season, year = 2017 for 2016-17 etc.).
}}
}
\description{
Allows to create a data frame for specific NBA team,
with their winning percentage
}
\examples{
teamdf('SAS', 2005)
}
